01-llm-eval-harness/requirements.txt
# Minimal set; pin as needed for your environment
transformers>=4.37.0
accelerate>=0.27.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24
scikit-learn>=1.3
pandas>=2.0
rank-bm25>=0.2.2
rich>=13.7
# Optional for fast inference; comment out if unnecessary
# vllm>=0.4.0
